requirements.txt
openai==1.3.5
Flask==3.0.0
python-dotenv==1.0.0